gta: fix not showing changed dependencies
This fixes an issue that was regressed with the latest changes. The
problem was tha 'prefixes', set by the '--include' flag, would filter in
the old version **after** the graph was computed. In the new version,
the filter would be applied during the computation, which would lead to
not pick anything.

This fixes the issue by moving the prefixes logic to the place where the
computation is already done.