tools/vendorlint: handle package removal
Teach Differ implementations (e.g. the Git type) to provide paths for
deleted files, too, so that clients can make fully informed decisions.
To solve the problem with vendorlinter, it would have been sufficient to
only modify Git.DiffFiles, but that would introduce an asymmetry between
Diff() and DiffFiles() that could surprise users, so both Diff() and
DiffFiles() are modified to return _all_ the changed paths and whether
or not they exist.