refactor to support modules
Refactor to support modules. I attempted to make the change as minimally
disruptive as possible so that all the existing code can still be
leveraged.

The Packager implementation now uses golang.org/x/tools/go/packages in
addition to go/build so that packages in modules can be resolved
correctly. It builds up a reverse dependency graph so that the canonical
import path can be determined when `go/build`'s calls to resolve
packages (i.e. Import or ImportDir) respond with a Package that uses a
relative import path