workaround issues with vendored packages in module aware mode
Work around issues with vendored packages in module aware mode. Since
module aware mode is the default as of Go 1.16 and Go 1.17 is scheduled
to drop support for GOPATH completely, this should be ok to do instead
of trying to be smart and figure out which mode to operate in.

We probably need to refactor the Packager implementation more fully to
completely quit relying on go/build.