load packages based on changed files
Load packages based on changed files, so that all the expected packages
will be in the forward graph. This is important when files for another
OS are changed and those files import packages that are not otherwise
imported by the files for the current build environment.

For example, given
* vcs.test/bar/foo_linux.go is changed
* vcs.test/bar/foo_linux.go imports vcs.test/quux
* no files in vcs.test/bar used to compile for Darwin import
  vcs.test/quux
* gta is run on darwin
* GOOS is not set

Then when gta tries to get the package for vcs.test/quux, it won't
exist, because "golang.org/x/tools/go/packages.Package".Load will not
have added any information about it.

Another alternative is to use "..." for the prefixes when calling
Package.Load, but that takes a bit longer.